Split loot handling code. Extend item set exports.
Common code is now located under `export.wiki.loot`.

Field `forceBP` renamed to `bpChance`: now a number between 0 and 1 (inclusive), with 0 being the default. `forceBP` carries as a `1` in the new field. Format version bump(!).

Removed unused `quality` field from ItemSet: was never exported from assets.

Added `canRepeatItems` to `ItemSet` model. Will address the naming inconsistency (with `noSetReplacements`) in a future commit. Another format bump, ouch! Sorry coldino!

Added fields requested by Tekcor: `rollOneItemOnly` (instead of once per quantity), `statMaxMult` (default 1, specifies how much a stat can exceed clamps on official servers or opted-in unofficial servers), `grindable` (default is false).
