+++ export/wiki/stage_missions.py
@@ -41,7 +41,7 @@ OUTPUT_FLAGS = (
 class MissionsStage(JsonHierarchyExportStage):
 
     def get_format_version(self) -> str:
-        return "2"
+        return "3"
 
     def get_name(self):
         return 'missions'
